requirements.txt
Flask==2.0.1
gunicorn==20.1.0
Werkzeug==2.0.3
requests==2.32.3
ipapi==1.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==7.4.0
Flask-Session==0.5.0
